requirements_x86.txt
numpy
onnx
pillow
pytest
